+++ src/Network/MPD/Commands/PlaybackControl.hs
@@ -20,6 +20,7 @@ module Network.MPD.Commands.PlaybackControl
     , previous
     , seek
     , seekId
+    , seekCur
     , stop
     ) where
 
@@ -55,6 +56,13 @@ seek pos = A.runCommand . A.seek pos
 -- | Seek to some point in a song (id version)
 seekId :: MonadMPD m => Id -> FractionalSeconds -> m ()
 seekId id' = A.runCommand . A.seekId id'
+
+-- | Seek to some point in the current song. Absolute time for True in
+-- the frist argument, relative time for False.
+--
+-- @since 0.9.2.0
+seekCur :: MonadMPD m => Bool -> FractionalSeconds -> m ()
+seekCur bool = A.runCommand . A.seekCur bool
 
 -- | Stop playing.
 stop :: MonadMPD m => m ()